Generic pmap.
