+++ sys/aarch64/pmap.c
@@ -1,38 +1,8 @@
 #define KL_LOG KL_PMAP
 #include <sys/klog.h>
-#include <sys/mimiker.h>
-#include <sys/libkern.h>
-#include <sys/pool.h>
-#include <aarch64/armreg.h>
-#include <aarch64/pte.h>
-#include <aarch64/tlb.h>
-#include <aarch64/pmap.h>
-#include <sys/pcpu.h>
 #include <sys/pmap.h>
-#include <sys/mutex.h>
-#include <sys/sched.h>
-#include <sys/vm_physmem.h>
-#include <bitstring.h>
-#include <sys/errno.h>
-#include <sys/kasan.h>
-
-typedef struct pmap {
-  mtx_t mtx;                      /* protects all fields in this structure */
-  asid_t asid;                    /* address space identifier */
-  paddr_t pde;                    /* directory page table physical address */
-  vm_pagelist_t pte_pages;        /* pages we allocate in page table */
-  TAILQ_HEAD(, pv_entry) pv_list; /* all pages mapped by this physical map */
-} pmap_t;
-
-typedef struct pv_entry {
-  TAILQ_ENTRY(pv_entry) pmap_link; /* link on pmap::pv_list */
-  TAILQ_ENTRY(pv_entry) page_link; /* link on vm_page::pv_list */
-  pmap_t *pmap;                    /* page is mapped in this pmap */
-  vaddr_t va;                      /* under this address */
-} pv_entry_t;
-
-static POOL_DEFINE(P_PMAP, "pmap", sizeof(pmap_t));
-static POOL_DEFINE(P_PV, "pv_entry", sizeof(pv_entry_t));
+#include <sys/_pmap.h>
+#include <aarch64/armreg.h>
 
 /*
  * This table describes which access bits need to be set in page table entry
@@ -76,162 +46,79 @@ static const pte_t vm_prot_map[] = {
     ATTR_AP_RW | ATTR_SW_READ | ATTR_SW_WRITE | ATTR_AF | pte_common,
 };
 
-static pmap_t kernel_pmap;
-paddr_t _kernel_pmap_pde;
-static bitstr_t asid_used[bitstr_size(MAX_ASID)] = {0};
-static SPIN_DEFINE(asid_lock, 0);
-
-/* this lock is used to protect the vm_page::pv_list field */
-/* the order of acquiring locks is as follows: firstly pv_list_lock and then
- * pmap_t::mtx */
-static MTX_DEFINE(pv_list_lock, 0);
-
-#define PAGE_OFFSET(x) ((x) & (PAGESIZE - 1))
-#define PG_DMAP_ADDR(pg) ((void *)((intptr_t)(pg)->paddr + DMAP_BASE))
+static pde_t pd_lvl_masks[] = {
+  [0] = L0_TABLE,
+  [1] = L1_TABLE,
+  [2] = L2_TABLE,
+};
 
 /*
- * Helper functions.
+ * Translation structure.
  */
-static bool user_addr_p(vaddr_t addr) {
-  return (addr >= USER_SPACE_BEGIN) && (addr < USER_SPACE_END);
+
+inline size_t l0_index(vaddr_t va) {
+  return L0_INDEX(va);
 }
 
-static bool kern_addr_p(vaddr_t addr) {
-  return (addr >= KERNEL_SPACE_BEGIN) && (addr < KERNEL_SPACE_END);
+inline size_t l1_index(vaddr_t va) {
+  return L1_INDEX(va);
 }
 
-inline vaddr_t pmap_start(pmap_t *pmap) {
-  return pmap->asid ? USER_SPACE_BEGIN : KERNEL_SPACE_BEGIN;
+inline size_t l2_index(vaddr_t va) {
+  return L2_INDEX(va);
 }
 
-inline vaddr_t pmap_end(pmap_t *pmap) {
-  return pmap->asid ? USER_SPACE_END : KERNEL_SPACE_END;
-}
-
-inline bool pmap_address_p(pmap_t *pmap, vaddr_t va) {
-  return pmap_start(pmap) <= va && va < pmap_end(pmap);
-}
-
-inline bool pmap_contains_p(pmap_t *pmap, vaddr_t start, vaddr_t end) {
-  return pmap_start(pmap) <= start && end <= pmap_end(pmap);
-}
-
-inline pmap_t *pmap_kernel(void) {
-  return &kernel_pmap;
-}
-
-inline pmap_t *pmap_user(void) {
-  return PCPU_GET(curpmap);
-}
-
-pmap_t *pmap_lookup(vaddr_t addr) {
-  if (kern_addr_p(addr))
-    return pmap_kernel();
-  if (user_addr_p(addr))
-    return pmap_user();
-  return NULL;
+inline size_t l3_index(vaddr_t va) {
+  return L3_INDEX(va);
 }
 
 /*
- * Address space identifiers management.
+ * Page directory.
  */
 
-static asid_t alloc_asid(void) {
-  int free = 0;
-  WITH_SPIN_LOCK (&asid_lock) {
-    bit_ffc(asid_used, MAX_ASID, &free);
-    if (free < 0)
-      panic("Out of asids!");
-    bit_set(asid_used, free);
+inline bool pde_valid_p(pde_t pde) {
+  return PTE_FRAME_ADDR(pde) != 0UL;
+}
+
+inline pde_t pde_make(paddr_t pa, unsigned lvl) {
+  assert(lvl < ADDR_TRANSLATION_DEPTH - 1);
+  return pa | pd_lvl_masks[lvl];
+}
+
+inline paddr_t pde2pa(pde_t pde) {
+  return PTE_FRAME_ADDR(pde);
+}
+
+void kernel_pd_change_notif(pmap_t *pmap, vaddr_t va, pde_t pde) {
+  /* Nothing to be done here. */
+}
+
+/*
+ * Page table.
+ */
+
+inline bool pte_valid_p(pte_t pte) {
+  return PTE_FRAME_ADDR(pte) != 0UL;
+}
+
+inline bool pte_readable(pte_t pte) {
+  return pte & ATTR_SW_READ;
+}
+
+inline bool pte_writable(pte_t pte) {
+  return pte & ATTR_SW_WRITE;
+}
+
+inline bool pte_executable(pte_t pte) {
+  return !(pte & ATTR_SW_NOEXEC);
+}
+
+pte_t pte_make(paddr_t pa, vm_prot_t prot, unsigned flags, bool kernel) {
+  pte_t pte = pa | vm_prot_map[prot];
+  if (!kernel) {
+    pte |= ATTR_AP_USER;
+    pte &= ~ATTR_AF;
   }
-  klog("alloc_asid() = %d", free);
-  return free;
-}
-
-static void free_asid(asid_t asid) {
-  klog("free_asid(%d)", asid);
-  SCOPED_SPIN_LOCK(&asid_lock);
-  bit_clear(asid_used, (unsigned)asid);
-  tlb_invalidate_asid(asid);
-}
-
-/*
- * Physical-to-virtual entries are managed for all pageable mappings.
- */
-
-static void pv_add(pmap_t *pmap, vaddr_t va, vm_page_t *pg) {
-  assert(mtx_owned(&pv_list_lock));
-  pv_entry_t *pv = pool_alloc(P_PV, M_ZERO);
-  pv->pmap = pmap;
-  pv->va = va;
-  TAILQ_INSERT_TAIL(&pg->pv_list, pv, page_link);
-  TAILQ_INSERT_TAIL(&pmap->pv_list, pv, pmap_link);
-}
-
-static pv_entry_t *pv_find(pmap_t *pmap, vaddr_t va, vm_page_t *pg) {
-  assert(mtx_owned(&pv_list_lock));
-  pv_entry_t *pv;
-  TAILQ_FOREACH (pv, &pg->pv_list, page_link) {
-    if (pv->pmap == pmap && pv->va == va)
-      return pv;
-  }
-  return NULL;
-}
-
-static void pv_remove(pmap_t *pmap, vaddr_t va, vm_page_t *pg) {
-  assert(mtx_owned(&pv_list_lock));
-  pv_entry_t *pv = pv_find(pmap, va, pg);
-  assert(pv != NULL);
-  TAILQ_REMOVE(&pg->pv_list, pv, page_link);
-  TAILQ_REMOVE(&pmap->pv_list, pv, pmap_link);
-  pool_free(P_PV, pv);
-}
-
-/*
- * Routines for accessing page table entries.
- */
-
-static vm_page_t *pmap_pagealloc(void) {
-  vm_page_t *pg = vm_page_alloc(1);
-  pmap_zero_page(pg);
-  return pg;
-}
-
-static pte_t *pmap_lookup_pte(pmap_t *pmap, vaddr_t va) {
-  pde_t *pdep;
-  paddr_t pa = pmap->pde;
-
-  /* Level 0 */
-  pdep = (pde_t *)PHYS_TO_DMAP(pa) + L0_INDEX(va);
-  if (!(pa = PTE_FRAME_ADDR(*pdep)))
-    return NULL;
-
-  /* Level 1 */
-  pdep = (pde_t *)PHYS_TO_DMAP(pa) + L1_INDEX(va);
-  if (!(pa = PTE_FRAME_ADDR(*pdep)))
-    return NULL;
-
-  /* Level 2 */
-  pdep = (pde_t *)PHYS_TO_DMAP(pa) + L2_INDEX(va);
-  if (!(pa = PTE_FRAME_ADDR(*pdep)))
-    return NULL;
-
-  /* Level 3 */
-  return (pde_t *)PHYS_TO_DMAP(pa) + L3_INDEX(va);
-}
-
-static paddr_t pmap_alloc_pde(pmap_t *pmap, vaddr_t vaddr) {
-  vm_page_t *pg = pmap_pagealloc();
-
-  TAILQ_INSERT_TAIL(&pmap->pte_pages, pg, pageq);
-
-  klog("Page table for 0x%016lx allocated at 0x%016lx", vaddr, pg->paddr);
-
-  return pg->paddr;
-}
-
-static pte_t make_pte(paddr_t pa, pte_t prot, unsigned flags) {
-  pte_t pte = pa | prot;
   unsigned cacheflags = flags & PMAP_CACHE_MASK;
   if (cacheflags == PMAP_NOCACHE)
     return pte | ATTR_IDX(ATTR_NORMAL_MEM_NC);
@@ -240,55 +127,22 @@ static pte_t make_pte(paddr_t pa, pte_t prot, unsigned flags) {
   return pte | ATTR_IDX(ATTR_NORMAL_MEM_WB);
 }
 
-static void pmap_write_pte(pmap_t *pmap, pte_t *ptep, pte_t pte, vaddr_t va) {
-  if (pmap != pmap_kernel())
-    pte |= ATTR_AP_USER;
-  *ptep = pte;
-  tlb_invalidate(va, pmap->asid);
+inline pte_t pte_protect(pte_t pte, vm_prot_t prot) {
+  return vm_prot_map[prot] | (pte & (~ATTR_AP_MASK & ~ATTR_XN));
+}
+
+inline paddr_t pte2pa(pte_t pte) {
+  return PTE_FRAME_ADDR(pte);
 }
 
 /*
- * Return pointer to entry of va in level 3 of page table. Allocate space if
- * needed.
+ * Physical map management.
  */
 
-static pte_t *pmap_ensure_pte(pmap_t *pmap, vaddr_t va) {
-  pde_t *pdep;
-  paddr_t pa = pmap->pde;
-
-  /* Level 0 */
-  pdep = (pde_t *)PHYS_TO_DMAP(pa) + L0_INDEX(va);
-  if (!(pa = PTE_FRAME_ADDR(*pdep))) {
-    pa = pmap_alloc_pde(pmap, va);
-    *pdep = pa | L0_TABLE;
-  }
-
-  /* Level 1 */
-  pdep = (pde_t *)PHYS_TO_DMAP(pa) + L1_INDEX(va);
-  if (!(pa = PTE_FRAME_ADDR(*pdep))) {
-    pa = pmap_alloc_pde(pmap, va);
-    *pdep = pa | L1_TABLE;
-  }
-
-  /* Level 2 */
-  pdep = (pde_t *)PHYS_TO_DMAP(pa) + L2_INDEX(va);
-  if (!(pa = PTE_FRAME_ADDR(*pdep))) {
-    pa = pmap_alloc_pde(pmap, va);
-    *pdep = pa | L2_TABLE;
-  }
-
-  /* Level 3 */
-  return (pde_t *)PHYS_TO_DMAP(pa) + L3_INDEX(va);
-}
-
-void pmap_activate(pmap_t *umap) {
-  SCOPED_NO_PREEMPTION();
-
-  PCPU_SET(curpmap, umap);
-
+void pmap_md_activate(pmap_t *umap) {
   uint64_t tcr = READ_SPECIALREG(TCR_EL1);
 
-  if (umap == NULL) {
+  if (!umap) {
     WRITE_SPECIALREG(TCR_EL1, tcr | TCR_EPD0);
   } else {
     uint64_t ttbr0 = ((uint64_t)umap->asid << ASID_SHIFT) | umap->pde;
@@ -297,331 +151,20 @@ void pmap_activate(pmap_t *umap) {
   }
 }
 
-/*
- * Wired memory interface.
- */
-
-void pmap_kenter(vaddr_t va, paddr_t pa, vm_prot_t prot, unsigned flags) {
-  pmap_t *pmap = pmap_kernel();
-
-  assert(page_aligned_p(pa) && page_aligned_p(va));
-  assert(pmap_address_p(pmap, va));
-  assert(pa != 0);
-
-  klog("Enter unmanaged mapping from %p to %p", va, pa);
-
-  pte_t pte = make_pte(pa, vm_prot_map[prot], flags);
-
-  WITH_MTX_LOCK (&pmap->mtx) {
-    pte_t *ptep = pmap_ensure_pte(pmap, va);
-    pmap_write_pte(pmap, ptep, pte, va);
-  }
+void pmap_md_setup(pmap_t *pmap __unused) {
+  /* Nothing to be done here. */
 }
 
-void pmap_kremove(vaddr_t va, size_t size) {
-  pmap_t *pmap = pmap_kernel();
-
-  assert(page_aligned_p(va) && page_aligned_p(size));
-  assert(pmap_contains_p(pmap, va, va + size));
-
-  klog("%s: remove unmanaged mapping for %p - %p range", __func__, va,
-       va + size - 1);
-
-  WITH_MTX_LOCK (&pmap->mtx) {
-    for (size_t off = 0; off < size; off += PAGESIZE) {
-      pte_t *ptep = pmap_lookup_pte(pmap, va + off);
-      assert(ptep != NULL);
-      pmap_write_pte(pmap, ptep, 0, va + off);
-    }
-  }
-}
-
-bool pmap_kextract(vaddr_t va, paddr_t *pap) {
-  return pmap_extract(pmap_kernel(), va, pap);
+void pmap_md_delete(pmap_t *pmap __unused) {
+  /* Nothing to be done here. */
 }
 
 /*
- * Pageable (user & kernel) memory interface.
+ * Bootstrap.
  */
 
-static bool pmap_extract_nolock(pmap_t *pmap, vaddr_t va, paddr_t *pap) {
-  if (!pmap_address_p(pmap, va))
-    return false;
-
-  pte_t *ptep = pmap_lookup_pte(pmap, va);
-  if (ptep == NULL)
-    return false;
-  paddr_t pa = PTE_FRAME_ADDR(*ptep);
-  if (pa == 0)
-    return false;
-  *pap = pa | PAGE_OFFSET(va);
-  return true;
-}
-
-void pmap_enter(pmap_t *pmap, vaddr_t va, vm_page_t *pg, vm_prot_t prot,
-                unsigned flags) {
-  paddr_t pa = pg->paddr;
-
-  assert(page_aligned_p(va));
-  assert(pmap_address_p(pmap, va));
-
-  klog("Enter virtual mapping %p for frame %p", va, pa);
-
-  bool kern_mapping = (pmap == pmap_kernel());
-
-  /* Mark user pages as non-referenced & non-modified. */
-  pte_t mask = kern_mapping ? 0UL : (ATTR_AF);
-  pte_t pte = make_pte(pa, vm_prot_map[prot] & ~mask, flags);
-
-  WITH_MTX_LOCK (&pv_list_lock) {
-    WITH_MTX_LOCK (&pmap->mtx) {
-      pv_entry_t *pv = pv_find(pmap, va, pg);
-      if (pv == NULL)
-        pv_add(pmap, va, pg);
-      if (kern_mapping)
-        pg->flags |= PG_MODIFIED | PG_REFERENCED;
-      else
-        pg->flags &= ~(PG_MODIFIED | PG_REFERENCED);
-      pte_t *ptep = pmap_ensure_pte(pmap, va);
-      pmap_write_pte(pmap, ptep, pte, va);
-    }
-  }
-}
-
-void pmap_remove(pmap_t *pmap, vaddr_t start, vaddr_t end) {
-  assert(page_aligned_p(start) && page_aligned_p(end) && start < end);
-  assert(pmap_contains_p(pmap, start, end));
-
-  klog("Remove page mapping for address range %p-%p", start, end);
-
-  WITH_MTX_LOCK (&pv_list_lock) {
-    WITH_MTX_LOCK (&pmap->mtx) {
-      for (vaddr_t va = start; va < end; va += PAGESIZE) {
-        pte_t *ptep = pmap_lookup_pte(pmap, va);
-        if (ptep == NULL)
-          continue;
-        paddr_t pa = PTE_FRAME_ADDR(*ptep);
-        if (pa == 0)
-          continue;
-        vm_page_t *pg = vm_page_find(pa);
-        pv_remove(pmap, va, pg);
-        pmap_write_pte(pmap, ptep, 0, va);
-      }
-    }
-  }
-}
-
-void pmap_protect(pmap_t *pmap, vaddr_t start, vaddr_t end, vm_prot_t prot) {
-  assert(page_aligned_p(start) && page_aligned_p(end) && start < end);
-  assert(pmap_contains_p(pmap, start, end));
-
-  klog("Change protection bits to %x for address range %p-%p", prot, start,
-       end);
-
-  WITH_MTX_LOCK (&pmap->mtx) {
-    for (vaddr_t va = start; va < end; va += PAGESIZE) {
-      pte_t *ptep = pmap_lookup_pte(pmap, va);
-      if (ptep == NULL)
-        continue;
-      pte_t pte = vm_prot_map[prot] | (*ptep & (~ATTR_AP_MASK & ~ATTR_XN));
-      pmap_write_pte(pmap, ptep, pte, va);
-    }
-  }
-}
-
-bool pmap_extract(pmap_t *pmap, vaddr_t va, paddr_t *pap) {
-  SCOPED_MTX_LOCK(&pmap->mtx);
-  return pmap_extract_nolock(pmap, va, pap);
-}
-
-void pmap_page_remove(vm_page_t *pg) {
-  SCOPED_MTX_LOCK(&pv_list_lock);
-
-  while (!TAILQ_EMPTY(&pg->pv_list)) {
-    pv_entry_t *pv = TAILQ_FIRST(&pg->pv_list);
-    pmap_t *pmap = pv->pmap;
-    vaddr_t va = pv->va;
-    WITH_MTX_LOCK (&pmap->mtx) {
-      TAILQ_REMOVE(&pg->pv_list, pv, page_link);
-      TAILQ_REMOVE(&pmap->pv_list, pv, pmap_link);
-      pte_t *ptep = pmap_lookup_pte(pmap, va);
-      assert(ptep != NULL);
-      pmap_write_pte(pmap, ptep, 0, va);
-    }
-    pool_free(P_PV, pv);
-  }
-}
-
-void pmap_zero_page(vm_page_t *pg) {
-  bzero(PG_DMAP_ADDR(pg), PAGESIZE);
-}
-
-void pmap_copy_page(vm_page_t *src, vm_page_t *dst) {
-  memcpy(PG_DMAP_ADDR(dst), PG_DMAP_ADDR(src), PAGESIZE);
-}
-
-static void pmap_modify_flags(vm_page_t *pg, pte_t set, pte_t clr) {
-  SCOPED_MTX_LOCK(&pv_list_lock);
-  pv_entry_t *pv;
-  TAILQ_FOREACH (pv, &pg->pv_list, page_link) {
-    pmap_t *pmap = pv->pmap;
-    vaddr_t va = pv->va;
-    WITH_MTX_LOCK (&pmap->mtx) {
-      pte_t *ptep = pmap_lookup_pte(pmap, va);
-      assert(ptep != NULL);
-      pte_t pte = *ptep;
-      pte |= set;
-      pte &= ~clr;
-      *ptep = pte;
-      tlb_invalidate(va, pmap->asid);
-    }
-  }
-}
-
-bool pmap_clear_referenced(vm_page_t *pg) {
-  bool prev = pmap_is_referenced(pg);
-  pg->flags &= ~PG_REFERENCED;
-  pmap_modify_flags(pg, 0, ATTR_AF);
-  return prev;
-}
-
-bool pmap_clear_modified(vm_page_t *pg) {
-  bool prev = pmap_is_modified(pg);
-  pg->flags &= ~PG_MODIFIED;
-  pmap_modify_flags(pg, ATTR_AP_RO, 0);
-  return prev;
-}
-
-bool pmap_is_referenced(vm_page_t *pg) {
-  return pg->flags & PG_REFERENCED;
-}
-
-bool pmap_is_modified(vm_page_t *pg) {
-  return pg->flags & PG_MODIFIED;
-}
-
-void pmap_set_referenced(vm_page_t *pg) {
-  pg->flags |= PG_REFERENCED;
-  pmap_modify_flags(pg, ATTR_AF, 0);
-}
-
-void pmap_set_modified(vm_page_t *pg) {
-  pg->flags |= PG_MODIFIED;
-  pmap_modify_flags(pg, 0, ATTR_AP_RO);
-}
-
-int pmap_emulate_bits(pmap_t *pmap, vaddr_t va, vm_prot_t prot) {
-  paddr_t pa;
-
-  WITH_MTX_LOCK (&pmap->mtx) {
-    if (!pmap_extract_nolock(pmap, va, &pa))
-      return EFAULT;
-
-    pte_t pte = *pmap_lookup_pte(pmap, va);
-
-    if ((prot & VM_PROT_READ) && !(pte & ATTR_SW_READ))
-      return EACCES;
-
-    if ((prot & VM_PROT_WRITE) && !(pte & ATTR_SW_WRITE))
-      return EACCES;
-
-    if ((prot & VM_PROT_EXEC) && (pte & ATTR_SW_NOEXEC))
-      return EACCES;
-  }
-
-  vm_page_t *pg = vm_page_find(pa);
-  assert(pg != NULL);
-
-  WITH_MTX_LOCK (&pv_list_lock) {
-    /* Kernel non-pageable memory? */
-    if (TAILQ_EMPTY(&pg->pv_list))
-      return EINVAL;
-  }
-
-  pmap_set_referenced(pg);
-  if (prot & VM_PROT_WRITE)
-    pmap_set_modified(pg);
-
-  return 0;
-}
-
-/*
- * Physical map management routines.
- */
-
-static void pmap_setup(pmap_t *pmap) {
-  pmap->asid = alloc_asid();
-  mtx_init(&pmap->mtx, 0);
-  TAILQ_INIT(&pmap->pte_pages);
-  TAILQ_INIT(&pmap->pv_list);
-}
-
-void init_pmap(void) {
-  pmap_setup(&kernel_pmap);
-  kernel_pmap.pde = _kernel_pmap_pde;
-}
-
-pmap_t *pmap_new(void) {
-  pmap_t *pmap = pool_alloc(P_PMAP, M_ZERO);
-  pmap_setup(pmap);
-
-  vm_page_t *pg = pmap_pagealloc();
-  TAILQ_INSERT_TAIL(&pmap->pte_pages, pg, pageq);
-  pmap->pde = pg->paddr;
-  klog("Page directory table allocated at %p", pmap->pde);
-
-  return pmap;
-}
-
-void pmap_delete(pmap_t *pmap) {
-  assert(pmap != pmap_kernel());
-
-  while (!TAILQ_EMPTY(&pmap->pv_list)) {
-    pv_entry_t *pv = TAILQ_FIRST(&pmap->pv_list);
-    vm_page_t *pg;
-    paddr_t pa;
-    pmap_extract_nolock(pmap, pv->va, &pa);
-    pg = vm_page_find(pa);
-    WITH_MTX_LOCK (&pv_list_lock)
-      TAILQ_REMOVE(&pg->pv_list, pv, page_link);
-    TAILQ_REMOVE(&pmap->pv_list, pv, pmap_link);
-    pool_free(P_PV, pv);
-  }
-
-  while (!TAILQ_EMPTY(&pmap->pte_pages)) {
-    vm_page_t *pg = TAILQ_FIRST(&pmap->pte_pages);
-    TAILQ_REMOVE(&pmap->pte_pages, pg, pageq);
-    vm_page_free(pg);
-  }
-
-  free_asid(pmap->asid);
-  pool_free(P_PMAP, pmap);
-}
-
-/*
- * Increase usable kernel virtual address space to at least maxkvaddr.
- * Allocate page table (levels 1, 2, 3) if needed.
- */
-void pmap_growkernel(vaddr_t maxkvaddr) {
-  assert(mtx_owned(&vm_kernel_end_lock));
-  assert(maxkvaddr > vm_kernel_end);
-
-  pmap_t *pmap = pmap_kernel();
-  vaddr_t va;
-
-  maxkvaddr = roundup2(maxkvaddr, L2_SIZE);
-
-  WITH_MTX_LOCK (&pmap->mtx) {
-    for (va = vm_kernel_end; va < maxkvaddr; va += L2_SIZE)
-      pmap_ensure_pte(pmap, va);
-  }
-
-  /*
-   * kasan_grow calls pmap_kenter which acquires pmap->mtx.
-   * But we are under vm_kernel_end_lock from kmem so it's safe to call
-   * kasan_grow.
-   */
-  kasan_grow(maxkvaddr);
-
-  vm_kernel_end = maxkvaddr;
+void pmap_bootstrap(paddr_t pd_pa) {
+  dmap_paddr_base = 0;
+  dmap_paddr_end = DMAP_SIZE;
+  kernel_pde = pd_pa;
 }
