+++ sys/gen/Makefile
@@ -1,0 +1,10 @@
+# vim: tabstop=8 shiftwidth=8 noexpandtab:
+
+TOPDIR = $(realpath ../..)
+
+SOURCES = \
+	pmap.c
+
+CPPFLAGS += -D_MACHDEP
+
+include $(TOPDIR)/build/build.kern.mk
